requirements.txt
requests
beautifulsoup4
selenium
pandas
pyyaml
